Overhaul Apps2Samsung Settingsview to allow settings for multiple apps
