+++ Jellyfin2Samsung-CrossOS/Info.plist
@@ -9,7 +9,7 @@
     <key>CFBundleIdentifier</key>
     <string>com.madebypatrick.apps2samsung</string>
     <key>CFBundleVersion</key>
-    <string>1.8.0-beta</string>
+    <string>2.5.1</string>
     <key>CFBundlePackageType</key>
     <string>APPL</string>
     <key>CFBundleSignature</key>
