+++ Jellyfin2Samsung-CrossOS/Views/AppSettingsView.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace Apps2Samsung.Views;
+
+public partial class AppSettingsView : UserControl
+{
+    public AppSettingsView()
+    {
+        InitializeComponent();
+    }
+}
