+++ Jellyfin2Samsung-CrossOS/Views/JellyfinSettingsView.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace Apps2Samsung.Views;
+
+public partial class JellyfinSettingsView : UserControl
+{
+    public JellyfinSettingsView()
+    {
+        InitializeComponent();
+    }
+}
