+++ Jellyfin2Samsung-CrossOS/Views/TvAppSettingsView.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace Apps2Samsung.Views;
+
+public partial class TvAppSettingsView : UserControl
+{
+    public TvAppSettingsView()
+    {
+        InitializeComponent();
+    }
+}
